feat(ui): split title and small changes
